add feature to delete data product

+++ progres-lanjutan/product-app/routes/web.php
@@ -31,7 +31,7 @@ Route::post('/product', [ProductController::class,'store'])->name("product-store
 Route::get('/product/{id}', [ProductController::class, 'show']);
 Route::get('/product/{id}/edit', [ProductController::class, 'edit'])->name('product-edit')->middleware('auth', 'RoleCheck:admin');
 Route::put('/product/{id}', [ProductController::class, 'update'])->name('product-update')->middleware('auth', 'RoleCheck:admin');
-Route::delete('/product/{id}',[ProductController::class,'destroy']);
+Route::delete('/product/{id}',[ProductController::class,'destroy'])->name('product-delete')->middleware('auth', 'RoleCheck:admin');
 
 Route::get('route_cont/{id}', [BarangController::class, 'index']);
